src/inventory/requirements.txt
functions-framework==3.*
google-api-core
google-cloud-discoveryengine
google-cloud-storage
google-cloud-pubsub
requests
